Remove Lecture1/exercise_solutions.ipynb
Although it is not a bad idea to have a notebook with solutions to the
exercises, it is not a good idea to have it as a separate file in the
same branch. It is better to have two permanent branches, one with and
the other without the solutions. That way it is possible to simply
merge any changes in the exercises to the solution notebook instead of
having to update the exercise texts in two separate notebooks manually.