Initialize lecture 2 (NumPy and Astropy) manual
The NumPy and Astropy sections that were previously in lecture 1 (Python
introduction) manual are now in the lecture 2 (NumPy) manual.

Matplotlib is thoroughly covered in lecture 3 and also briefly covered
in the Dynamical Astronomy course materials. The Matplotlib section in
lecture 1 manual was therefore not needed and has been deleted.