Update the script for generating title slides
Daniel Mikkola is no longer involved in teaching this course, so he is
not mentioned on the title slides.

The generated files are now saved as `.jpeg` instead of `.png` because
of smaller file size.

The font used on the title slides is no longer specified explicitly
because the previously used font might not be present on all computers.