Improve formatting in `numpy` lesson manual
`black` inserts line breaks at trailing commas even if the line in
question does not exceed the 88 character limit. This caused a couple of
instances of bad formatting in the `numpy` lesson manual. There was also
a case of a line just exceeding the length limit, which has been
refactored. Unrelated to `black` the phrasing in one of the lines has
been updated.