Update script for generating title slides
Simon Alinder is now the only teacher of this course, so Eero was
removed from the list, and title slides were regenerated.

The name of the organization has changed from "Department of Astronomy
and Theoretical Particle Physics" to "Division of Astrophysics" since
the last time the course was given.

It seems some changes were done automatically to the bash script.