Update script for generating title slides
Simon Alinder is now the only teacher of this course, so Eero was
removed from the list, and title slides were regenerated.

The name of the organization has changed from "Department of Astronomy
and Theoretical Particle Physics" to "Division of Astrophysics" since
the last time the course was given.

Changed the front page generator script to use Liberation font. This
way, we ensure readability because the default font may be bad.
Because this increases complexity and creates the danger that future
users will not have this font, a README was added with some
explanations.

It seems some changes were done automatically to the bash script.