Update website by removing Eero and old link
There was a download link to spyderexample.py, which does not exist
anymore. This link was removed.

Revert deletion of xy.txt.
Turns out, it is not redundant after all, it is used in manual 4.
This reverts commit 2a1c2b0181df899b509f7b874dff237500ffe309.

Simon is the only teacher now so the presentation of Eero was removed
from the index page.

Added list of FOSS projects Simon has contributed to.