Add part in presentation 1 about Python versions
There has been significant changes in Python recently, in particular
regarding error messages. The newer ones are much more helpful which
should have a large impact on how inexperienced students interact with
Python. This also includes a new image taken from the Python webpage.

Also added a note that Astropy needs Python 3.9. This is also a natural
place to mention the troubles with Astropy and Anaconda.