Strongly reduce the scope of ruff linting
The code is in good shape but needs to include several examples of bad
or alternative ways of doing things.
Putting in ignores everywhere it would be needed would be distracting in
the presentations and provide a greater workload for the (single)
maintainer as the rules are frequently being updated.

Ideally, the "F" section (pyflakes) should be included but it literally
only produces false positives on the current code.
"I" (isort) and "Q" (flake8-quotes) are similar. I would like to include
them but they cause too much trouble with the notebook format and
intentionally different code.